requirements.txt
django-restframework
django-colorfield